+++ src/assertions/__tests__/toContainReact--tests.js
@@ -22,19 +22,19 @@ function Fixture() {
   );
 }
 
-describe('toContain', () => {
+describe('toContainReact', () => {
   it('works with `shallow` renders', () => {
     const wrapper = shallow(<Fixture />);
-    expect(wrapper).toContain(<User index={1} />);
+    expect(wrapper).toContainReact(<User index={1} />);
   });
 
   it('works with `mount` renders', () => {
     const wrapper = mount(<Fixture />);
-    expect(wrapper).toContain(<User index={1} />);
+    expect(wrapper).toContainReact(<User index={1} />);
   });
 
   it('works with with jasmines negation', () => {
     const wrapper = shallow(<Fixture />);
-    expect(wrapper).not.toContain(<User index={3} />);
+    expect(wrapper).not.toContainReact(<User index={3} />);
   });
 });