+++ src/assertions/__tests__/toMatchSelector--tests.js
@@ -9,25 +9,25 @@ function Fixture() {
   );
 }
 
-describe('toContain', () => {
+describe('toMatchSelector', () => {
   it('works with `shallow` renders', () => {
     const wrapper = shallow(<Fixture />);
 
-    expect(wrapper.find('#child')).toMatch('span');
-    expect(wrapper.find('#child')).toMatch('#child');
-    expect(wrapper.find('#child')).toMatch('.foo');
+    expect(wrapper.find('#child')).toMatchSelector('span');
+    expect(wrapper.find('#child')).toMatchSelector('#child');
+    expect(wrapper.find('#child')).toMatchSelector('.foo');
   });
 
   it('works with `mount` renders', () => {
     const wrapper = mount(<Fixture />);
 
-    expect(wrapper.find('#child')).toMatch('span');
-    expect(wrapper.find('#child')).toMatch('#child');
-    expect(wrapper.find('#child')).toMatch('.foo');
+    expect(wrapper.find('#child')).toMatchSelector('span');
+    expect(wrapper.find('#child')).toMatchSelector('#child');
+    expect(wrapper.find('#child')).toMatchSelector('.foo');
   });
 
   it('works with with jasmines negation', () => {
     const wrapper = shallow(<Fixture />);
-    expect(wrapper.find('#child')).not.toMatch('ball');
+    expect(wrapper.find('#child')).not.toMatchSelector('ball');
   });
 });