+++ src/assertions/toContainReact.js
@@ -2,14 +2,14 @@
  * This source code is licensed under the MIT-style license found in the
  * LICENSE file in the root directory of this source tree. *
  *
- * @providesModule toContain
+ * @providesModule toContainReact
  * @flow
  */
 
 import { shallow } from 'enzyme';
 
 export default {
-  toContain() : Object {
+  toContainReact() : Object {
     return {
       compare(enzymeWrapper:Object, reactInstance:Object) : Object {
         const wrappedInstance = shallow(reactInstance);